% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csStat_H0_func.R
\name{csStat_H0_func}
\alias{csStat_H0_func}
\title{Cumulative mean values of independently generated p-values under the null hypothesis.}
\usage{
csStat_H0_func(m, N.sim)
}
\arguments{
\item{m}{Number of individual test statistics (hypothesis tests) in the global test.}

\item{N.sim}{Number of simulations for calculation of cumulative mean
values of generated p-values.}
}
\value{
A matrix with the dimension \code{m*N.sim} containing the cumulative mean values
of individual p-values which are possibly transformed and then sorted.
Each column corresponds to a simulation.
}
\description{
This function generates the independent p-values under the null
from the uniform distribution. It then calculates cumulative means of
transformed sorted p-values to be used in the implementation of the omnibus test.
The transformation as \eqn{h(p)=1/p} was considered to give weight to extreme p-values.
}
\references{
Futschik, A., Taus, T., & Zehetmayer, S. (2019). An omnibus test for the global
null hypothesis. \emph{Statistical methods in medical research}, \strong{28}(8), 2292-2304.
}
\seealso{
\code{\link{omnibus_test}}
}
\author{
Sonja Zehetmayer
}
