% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_funcs.R
\name{est_funcs}
\alias{est_funcs}
\title{Estimation of an Item Response Theory (IRT) model, a linear model of
items and the corresponding latent traits.}
\usage{
est_funcs(longdata)
}
\arguments{
\item{longdata}{A data frame containing repeated measures of patients.
This dataset is used for estimation of the IRT model, linear model of items,
and corresponding latent trait estimation. The variables include:
\describe{
\item{\code{ID}}{Unique patient identifier.}
\item{\code{STUDY_ID}}{1, referring to the ABBV-8E12 trial.}
\item{\code{AGE}}{Age of patients.}
\item{\code{Sex}}{1 for Female, 2 for Male.}
\item{\code{PHENOTYPE}}{PSP-RS = 9, PSP diagnostic phenotype.}
\item{\code{TRT}}{Treatment label; 1 = test group, 0 = control group.}
\item{\code{VISIT}}{Visit number; can be 1, 6, 11, 14, 18
(\code{VISIT = 1} baseline; \code{VISIT = 18} week 52).}
\item{\code{TIME}}{Time in years.}
\item{\code{DV}}{Scores of the 10-item PSPRS recommended by FDA (mostly 0–4).}
\item{\code{ITEM}}{Item index for the FDA 10-item PSPRS. Originally numbered
3, 4, 5, 12, 13, 24, 25, 26, 27, 28; here renamed 1–10.}
\item{\code{DOMAIN}}{Domain number.}
\item{\code{SCALE}}{Global scale (PSPRS = 1).}
\item{\code{RICH}}{Indicator; 1 if \code{PHENOTYPE == 9}, otherwise 0.}
\item{\code{INTERV}}{1 if the study is interventional.}
}}
}
\value{
A list containing various model estimates (values with suffix
\code{.RESC} are based on rescored items):
\describe{
\item{\code{IRTmodel.full}}{Full IRT model.}
\item{\code{IRTpar.est}}{Estimated parameters for the IRT model.}
\item{\code{Linmodel.PSIpf}}{Linear model of items with the (transformed)
IRT latent trait as the dependent variable.}
\item{\code{est.IRTlatent.pf}}{Estimated latent traits for each patient from the IRT model.}
\item{\code{est.LMlatent.bpf}}{Approximated latent traits for each patient from the linear model of items.}
\item{\code{id.overlap}}{Patient IDs with item scores at both baseline and week 52.}
}
}
\description{
This function receives a simulated data set in the long format as the input.
The input simulated data set resembles items scores of the Progressive
Supranuclear Palsy (PSP) as in the ABBV-8E12 trial. The simulated data set
must contain specific variables as explained in data description part, below.
After accumulating data from independent time points (visits, \code{VISIT})
with the aim to enhance the precision of the model estimates, the graded
response IRT model is estimated. Using the estimated latent trait (progression rate)
for each patient from the IRT model, as the dependent variable, the function
fits a linear model with the 10 FDA recommended items as independent variables.
This yields an approximation of the latent variable based on a weighted sum
of the item scores. This novel approach for approximation of the latent trait
is simpler to compute without going through complex derivations. Similar
estimations are performed after the items are rescored based on the FDA suggestion.
In addition to the above estimations, the function also returns the patient IDs
who were present at both time points of interest (baseline and week 52
corresponding to \code{VISIT=1}, \code{VISIT=18}).
}
\details{
The input scores include the (simulated) 10-item version of the PSPRS recommended by FDA.
The item scores are categorical values which mostly vary from zero to 4. The estimations based on the
true data set of the ABBV-8E12 trial are presented in Yousefi et al. (2023).
}
\examples{
# Example usage:
\dontrun{
sim_data <- read.csv("C:/2024/PSP_directory/MultiendPSP/effMultiend/inst/extdata/simPSP_lim.csv")
result <- est_funcs(sim_data)
}

}
\references{
Ueckert, S. (2018). Modeling composite assessment data using item response theory.
\emph{CPT: Pharmacometrics & Systems Pharmacology}, \strong{7}(4), 205-218.
Yousefi, E., Gewily, M., K\"{o}nig, F., H\"{o}glinger, G., Hopfner, F.,
Karlsson, M. O., Ristl, R., Zehetmayer, S. & Posch, M. (2023).
Efficiency of Multivariate Tests in Trials in Progressive Supranuclear Palsy.
\emph{arXiv preprint arXiv:2312.08169}.
}
\seealso{
\code{\link{simPSPdata}}
}
\author{
Elham Yousefi
}
