% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omnibus_test.R
\name{omnibus_test}
\alias{omnibus_test}
\title{The Omnibus test for testing the global null hypothesis based on independent p-values}
\usage{
omnibus_test(p, csStat.H00)
}
\arguments{
\item{p}{A vector of p-values.}

\item{csStat.H00}{Null distribution of cumulative mean values.}
}
\value{
The omnibus test statistic.
}
\description{
The function calculates an Omnibus test statistic using the given independent
p-values and the null distribution of cumulative mean values. It focuses on
a general situation where independent p-values are available from several hypothesis tests that
are assumed to be uniformly distributed under the null hypothesis. It is an
alternative test for the global null hypothesis of no treatment effect in any
item which is based on cumulative sums of (possibly) transformed sorted p-values.
}
\references{
Futschik, A., Taus, T., & Zehetmayer, S. (2019). An omnibus test for the
global null hypothesis. \emph{Statistical methods in medical research}, \strong{28}(8), 2292-2304.
}
\seealso{
\code{\link{csStat_H0_func}}
}
\author{
Sonja Zehetmayer
}
