% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_IRTsim.R
\name{plot_IRTsim}
\alias{plot_IRTsim}
\title{Plot of results of analysis tests of the IRT-based simulations}
\usage{
plot_IRTsim(
  simresult,
  effectRatio = NULL,
  alphaa = 0.025,
  test.names = c("IRT.PSIF", "LM.PSIBPF", "SumS", "OLS", "GLS", "GLS_26", "Bonf", "Tmin",
    "Simes", "Omnibus", "Omnibus_dom"),
  scoretype = c("Original", "Rescore")
)
}
\arguments{
\item{simresult}{Result of simulations for both original and rescored
simulated items. See \code{\link{sim_run_par}} for more details. This include the following items:
\itemize{
\item \code{final.res.ORIG}: refer to returned values of the \link{sim_run_par} function for more details. This part of results can be accessed using the dollar operator as simresult$final.res.ORIG.
\item \code{final.res.RESC}: refer to returned values of the \link{sim_run_par} function for more details. This part of results can be accessed using the dollar operator as simresult$final.res.RESC.
}}

\item{effectRatio}{Vector of the rho values, indicating a beneficial
treatment effect, to be used in power simulations. It is used to model disease
progression under the alternative. See section 4.1 of the following paper.}

\item{alphaa}{Alpha level (default is 0.025).}

\item{test.names}{refer to the \link{sim_run_par} function for more details.}

\item{scoretype}{A character vector specifying type of scores. \code{scoretype="Original"} refers to the results when the original scaling provided by the FDA has been used.
\code{scoretype="Rescore"} refers to the case with the FDA rescoring. See the referenced paper below for more details.}
}
\value{
Plot of the simulation results.
}
\description{
This function plots the results of multiple analysis tests derived from the
parallel simulations described in \code{\link{sim_run_par}} and
\code{\link{test_fun_IRT}}. If \code{effectRatio.=NULL} the estimated
type one error rate is plotted. Otherwise, based on the values of
\code{effectRatio=unlist(effectRatio.list)} result of the power simulations
of analysis tests is plotted.
}
\examples{
# Example usage:
\dontrun{
Dvec <- seq(0.45,0.75,by=0.05) # vector
sim_data <- read.csv(
"C:/2024/PSP_directory/MultiendPSP/effMultiend/inst/extdata/simPSP_lim.csv"
)
csStat.H0item <- csStat_H0_func(m = 10, N.sim = 100000)
csStat.H0domain <-csStat_H0_func(m=3,N.sim=100000)
simresult <- sim_run_par(ni = 50, nj = 50, dat = sim_data,
effectRatio.list = as.list(Dvec), n.sim = 10000,
test.names=c("IRT.PSIF", "LM.PSIBPF", "SumS",  "OLS", "GLS", "GLS_26",
 "Bonf", "Tmin", "Simes", "Omnibus", "Omnibus_dom"),
csStat.H0=csStat.H0item, csStat.H03dom=csStat.H0domain)


plot_IRTsim(simresult=simresult, effectRatio = effectRatio, alphaa = 0.025,
test.names=c("IRT.PSIF", "LM.PSIBPF", "SumS",  "OLS", "GLS", "GLS_26",
 "Bonf", "Tmin", "Simes", "Omnibus", "Omnibus_dom"),
scoretype=c("Original","Rescore"))
}

}
\references{
Yousefi, E., Gewily, M., K\"{o}nig, F., H\"{o}glinger, G., Hopfner, F.,
Karlsson, M. O., Ristl, R., Zehetmayer, S. & Posch, M. (2023).
Efficiency of Multivariate Tests in Trials in Progressive Supranuclear Palsy.
\emph{arXiv preprint arXiv:2312.08169}.
}
\seealso{
\code{\link{simPSPdata}},
\code{\link{test_fun_IRT}} for details on the underlying analysis methods and
\code{\link{sim_run_par}}.
}
\author{
Elham Yousefi
}
