% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_run_par.R
\name{sim_run_par}
\alias{sim_run_par}
\title{parallel run of the IRT_based Simulations}
\usage{
sim_run_par(
  ni,
  nj,
  dat,
  effectRatio.list = NULL,
  alphaa = 0.025,
  usefitted = TRUE,
  n.sim,
  test_names = c("IRT.PSIF", "LM.PSIBPF", "SumS", "OLS", "GLS", "GLS_26", "Bonf", "Tmin",
    "Simes", "Omnibus", "Omnibus_dom")
)
}
\arguments{
\item{ni}{Number of subjects in the test group.}

\item{nj}{Number of subjects in the control group.}

\item{dat}{A data set in the long format, containing the necessary variables
as in \code{\link{simPSPdata}} data set to perform the simulations.}

\item{effectRatio.list}{A list of numeric values or NULL. A list of the rho
values, indicating a beneficial treatment effect, to be used in power
simulations. It is used to model disease progression under the alternative.
If NULL, simulations are performed to estimate the type I error rate,
otherwise the power simulations are performed. Default = NULL.}

\item{alphaa}{Alpha level. Default is 0.025.}

\item{usefitted}{Logical; if TRUE, use fitted values in the simulation. Default = TRUE.}

\item{n.sim}{Number of simulation runs.}

\item{test_names}{A character vector specifying the names of tests to be performed. This includes all the following tests:
\itemize{
\item \code{"IRT.PSIF"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"LM.PSIBPF"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"SumS"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"OLS"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"GLS"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"GLS_26"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"Bonf"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"Tmin"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"Simes"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"Omnibus"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
\item \code{"Omnibus_dom"}: refer to the documentation of the \link{test_fun_IRT} function for more details.
}}
}
\value{
A data frame containing the results of simulations for all the
analysis tests i.e., the proportions of time the H0 has been rejected.
This includes the results for the original and rescored simulated samples as the following:
\item{final.res.ORIG}{Results of simulations of all analysis tests
for the simulated items with the original scores}
\item{final.res.RESC}{Results of simulations of all analysis tests
for the rescored simulated items}
}
\description{
This function runs in parallel the IRT based simulations for estimation of
the type one error rate and simulation of the power of statistical hypothesis
tests. The analysis methods are described in \code{\link{test_fun_IRT}}.
If \code{effectRatio.list=NULL} the simulations are run to estimate the type one error rate.
}
\examples{
library(furrr)
library(parallel)
library(dplyr)
library(tidyr)
library(future)
library(matrixStats)
library(mirt)
library(dplyr)
library(tidyr)
library(tidyselect)
library(Matrix)
library(hommel)
library(multcomp)
library(mvtnorm)
\dontrun{
# Example usage:
Dvec.list <- as.list(seq(0.45,0.75,by=0.05))
sim_data <- read.csv(
"C:/2024/PSP_directory/MultiendPSP/effMultiend/inst/extdata/simPSP_lim.csv"
)
set.seed(2403)
csStat.H0item <- csStat_H0_func(m = 10, N.sim = 100000)
set.seed(2310)
csStat.H0domain <-csStat_H0_func(m=3,N.sim=100000)
analys_sim_result <- sim_run_par(ni = 50, nj = 50, dat = sim_data,
effectRatio.list = Dvec.list, n.sim = 10000,
test_names=c("IRT.PSIF", "LM.PSIBPF", "SumS",  "OLS", "GLS", "GLS_26",
 "Bonf", "Tmin", "Simes", "Omnibus", "Omnibus_dom"))
}

}
\references{
Yousefi, E., Gewily, M., K\"{o}nig, F., H\"{o}glinger, G., Hopfner, F.,
Karlsson, M. O., Ristl, R., Zehetmayer, S. & Posch, M. (2023).
Efficiency of Multivariate Tests in Trials in Progressive Supranuclear Palsy.
\emph{arXiv preprint arXiv:2312.08169}.
}
\seealso{
\code{\link{simPSPdata}}
\code{\link{test_fun_IRT}} for details on the underlying analysis methods.
\code{\link{csStat_H0_func}}
\code{\link{omnibus_test}}
}
\author{
Elham Yousefi
}
