% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPSPdata_read.R
\name{simPSPdata_read}
\alias{simPSPdata_read}
\title{Read CSV duplicate}
\usage{
simPSPdata_read(path)
}
\arguments{
\item{path}{path to file name}
}
\value{
a \code{data.frame}
}
\description{
Read CSV duplicate
}
\examples{
CSV = system.file("extdata", "simPSPdata.csv", package= "effMultiend")
simPSPdata_read(CSV)
}
