% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fun_IRT.R
\name{test_fun_IRT}
\alias{test_fun_IRT}
\title{IRT-based data generation of 10-item version of the PSPRS and performing
multiple analysis tests for multivariate endpoints}
\usage{
test_fun_IRT(ni, nj, dat, effectRatio = NULL, alphaa = 0.025, usefitted = TRUE)
}
\arguments{
\item{ni}{Numeric. Sample size for the test group. It should satisfy the
condition ni + nj <= length(unique(dat$ID)).}

\item{nj}{Numeric. Sample size for the control group. It should satisfy the
condition ni + nj <= length(unique(dat$ID)).}

\item{dat}{A data set in the long format, containing the necessary variables
as in \code{\link{simPSPdata}} data set to generate the IRT-based data from
and perform the analysis test for the generated data.}

\item{effectRatio}{A rho value, indicating a beneficial treatment effect,
for estimation of the power. It is used to model disease progression under
the alternative. If NULL the tests are performed under the null.}

\item{alphaa}{Numeric. Significance level for the test.}
}
\value{
A list of results of analysis tests based on both the original
simulated item scores and the rescored ones. The results with the extension
'.RESC' stands for the rescored values and holds the same interpretation as
those with the original scores, with the following components:
\item{pval.OLSij}{p-value of the OLS test}
\item{res.OLSij}{Indicator variable which takes the value of 1,
if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.GLSij}{p-value of the GLS test}
\item{res.GLSij}{Indicator variable which takes the value of 1,
if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.bonfij}{p-value of the Bonferroni-corrected test}
\item{res.bonfij}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.tmax}{p-value for the T.max analysis test}
\item{res.tmaxij}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.SumScoreij}{p-value for the SumScore analysis test}
\item{res.SumScoreij}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.Simesij}{p-value for the Simes analysis test}
\item{res.Simesij}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.Omnibus}{p-value for the Omnibus test}
\item{res.Omnibusij}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.GLSij26}{p-value for the GLS analysis test, without item 26}
\item{res.GLSij26}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{df.residual}{Degrees of freedom for residuals.}
\item{pval.PSIfij}{p-value for the 'IRT.PSIF' test}
\item{res.PSIfij}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.linmPSIbpfij}{p-value for the 'LM.PSIBPF' test}
\item{res.linmPSIbpfij}{1, if the test is significant (p-value < alphaa), otherwise 0}
\item{pval.Omnibus3Domij}{p-value for the Omnibus test across 3 domains ('Omnibus-dom')}
\item{res.Omnibus3Domij}{1, if the test is significant (p-value < alphaa), otherwise 0}
}
\description{
The function first generates IRT-model-based item level data for the baseline
and Week 52 item scores based on a longitudinal IRT model fitted on multiple
data sets from a set of interventional trials in PSP (See  Gewily et al. (2023)
and Section 4.1 of Yousefi et al. (2023) for details).
The function then performs various analysis tests as described in Section 3.2
of Yousefi et al. (2023). The analysis methods include three tests based on
the univariate test statistics (of aggregated scores) which test the global null
hypothesis that the treatments do not differ in any of the individual endpoints.
Sum Score test ('SumS'), the IRT-based test ('IRT.PSIF'), the approximate
IRT-based test ('LM.PSIBPF') are three analysis tests in this category, where
'IRT.PSIF' and 'LM.PSIBPF' are novel tests based on IRT. The O'Brien's
Ordinary Least Squares ('OLS') and Generalised Least Squares ('GLS') directional
tests applied to multivariate endpoints (items) at baseline and week 52 are
weighted test statistics composed of that of the individual items. We estimated
individual test statistics of items using multiple marginal models.
A marginal ANCOVA model is fitted to simulated scores of each item where the
simulated item score at week 52 is used as the dependent variable, and the
treatment variable and the simulated item score at baseline are used as the
independent variables. The analysis tests include implementation of other
standard tests based on individual components of the multivariate endpoint,
where the inference is based on the individual test statistics or the
individual p-values for the multiple endpoints and a multiplicity correction
is applied to control the familywise error rate in the strong sense.
Bonferroni-corrected test ('Bonf'), the test based on the maximum T-value
('MaxT'), Simes test ('Simes'), Omnibus test ('Omnibus'), and a combination
test based on Omnibus and Sum Score test ('Omnibus-dom') are examples of this
category. The same tests are applied to the simulated item scores after rescoring.
}
\examples{
\dontrun{
# Example usage:
# The value \rho < 1 indicates a beneficial treatment effect
# for building progression model in power simulations
Dval <- 0.45
sim_data <- sim_data <- read.csv(
"C:/2024/PSP_directory/MultiendPSP/effMultiend/inst/extdata/simPSP_lim.csv"
)
csStat.H0item <- csStat_H0_func(m = 10, N.sim = 100000)
csStat.H0domain <-csStat_H0_func(m=3,N.sim=100000)
analys_test_result <- test_fun_IRT(
ni = 70,
nj = 70,
dat = sim_data,
effectRatio = Dval,
alphaa = 0.025,
usefitted=TRUE)
}

}
\references{
Yousefi, E., Gewily, M., K\"{o}nig, F., H\"{o}glinger, G., Hopfner, F.,
Karlsson, M. O., Ristl, R., Zehetmayer, S. & Posch, M. (2023).
Efficiency of Multivariate Tests in Trials in Progressive Supranuclear Palsy.
\emph{arXiv preprint arXiv:2312.08169}.
}
\seealso{
\code{\link{simPSPdata}}
\code{\link{csStat_H0_func}}
\code{\link{omnibus_test}}
\code{\link{est_funcs}}
}
\author{
Elham Yousefi
}
\keyword{internal}
