% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_funcs.R
\name{est_funcs}
\alias{est_funcs}
\title{Estimation of an Item Response Theory (IRT) model, a linear model of
items and the corresponding latent traits.}
\usage{
est_funcs(longdata)
}
\arguments{
\item{longdata}{A data frame containing repeated measures of patients.
This dataset is used for estimation of the IRT model, linear model of items,
and corresponding latent trait estimation.
The variables include:
- \item{ID}{Unique patient identifier}
- \item{STUDY_ID}{1, referring to the ABBV-8E12 trial}
- \item{AGE}{Age of patients}
- \item{Sex}{1 for Female, 2 for Male}
- \item{PHENOTYPE}{PSP-RS ~ 9, PSP Diagnostic phenotype}
- \item{TRT}{Treatment, simulated treatment labels, 1 corresponds to the test group and 0 to the control group}
- \item{VISIT}{Visit number, where VISIT can take the values 1, 6, 11, 14, 18. VISIT=1 and VISIT=18 correspond to baseline and week 52}
- \item{TIME}{Time in year}
- \item{DV}{Dependent variable, scores of the 10-item version of the PSPRS recommended by FDA, mostly vary from 0 to 4}
- \item{ITEM}{Item measured, representing the 10-item version of the PSPRS recommended by the FDA. Originally, these items were numbered 3, 4, 5, 12, 13, 24, 25, 26, 27, 28. In our implementation, we renamed them from 1 to 10}
- \item{DOMAIN}{Domain number}
- \item{SCALE}{PSPRS ~ 1, Global scale}
- \item{RICH}{Indicator variable, taking the value 1 if PHENOTYPE is 9, otherwise 0}
- \item{INTERV}{1, if the study has been intervention}}
}
\value{
A list containing various model estimates with the following components.
Note that values with the extension '.RESC' stand for the rescored items and have similar interpretations as those in the following:
\item{IRTmodel.full}{Full IRT model}
\item{IRTpar.est}{Estimated parameters for the IRT model.}
\item{Linmodel.PSIpf}{Linear model of items with the (transformed) estimated
latent trait from the IRT model as the depedent variable}
\item{est.IRTlatent.pf}{Estimated latent traits for each patient from the IRT model}
\item{est.LMlatent.bpf}{Approximated latent traits for each patient from the linear model of items}
\item{id.overlap}{Patient IDs who have all item scores both at baseline and week 52}
}
\description{
This function receives a simulated data set in the long format as the input.
The input simulated data set resembles items scores of the Progressive
Supranuclear Palsy (PSP) as in the ABBV-8E12 trial. The simulated data set
must contain specific variables as explained in data description part, below.
After accumulating data from independent time points (visits, \code{VISIT})
with the aim to enhance the precision of the model estimates, the graded
response IRT model is estimated. Using the estimated latent trait (progression rate)
for each patient from the IRT model, as the dependent variable, the function
fits a linear model with the 10 FDA recommended items as independent variables.
This yields an approximation of the latent variable based on a weighted sum
of the item scores. This novel approach for approximation of the latent trait
is simpler to compute without going through complex derivations. Similar
estimations are performed after the items are rescored based on the FDA suggestion.
In addition to the above estimations, the function also returns the patient IDs
who were present at both time points of interest (baseline and week 52
corresponding to \code{VISIT=1}, \code{VISIT=18}).
}
\details{
The input scores include the (simulated) 10-item version of the PSPRS recommended by FDA.
The item scores are categorical values which mostly vary from zero to 4. The estimations based on the
true data set of the ABBV-8E12 trial are presented in Yousefi et al. (2023).
}
\examples{
# Example usage:
\dontrun{
sim_data <- read.csv("C:/2024/PSP_directory/MultiendPSP/effMultiend/inst/extdata/simPSP_lim.csv")
result <- est_funcs(sim_data)
}

}
\references{
Ueckert, S. (2018). Modeling composite assessment data using item response theory.
\emph{CPT: Pharmacometrics & Systems Pharmacology}, \strong{7}(4), 205-218.
Yousefi, E., Gewily, M., K\"{o}nig, F., H\"{o}glinger, G., Hopfner, F.,
Karlsson, M. O., Ristl, R., Zehetmayer, S. & Posch, M. (2023).
Efficiency of Multivariate Tests in Trials in Progressive Supranuclear Palsy.
\emph{arXiv preprint arXiv:2312.08169}.
}
\seealso{
\code{\link{simPSPdata}}
}
\author{
Elham Yousefi
}
